Final Commit
Set a timeout function to stop showing hint after 0.5 seconds, made the background of table cells change if color is wrong so signify a mistake, created an alert to tell the user when the puzzle is correct, and created an alert to tell the user they will lose progress if they change puzzles. Ready for review.